requirements.txt
ebaysdk
pytrends
requests
python-dotenv
beautifulsoup4
huggingface_hub
fake-useragent 
